+++ packages/adblocker-webextension/adblocker.ts
@@ -8,13 +8,21 @@
 
 import { parse } from 'tldts-experimental';
 
-import { FiltersEngine, Request, WebRequestType } from '@cliqz/adblocker';
+import {
+  FiltersEngine,
+  HTMLSelector,
+  isUTF8,
+  Request,
+  StreamingHtmlFilter,
+  WebRequestType,
+} from '@cliqz/adblocker';
 import { IBackgroundCallback, IMessageFromBackground } from '@cliqz/adblocker-content';
 
 export interface WebRequestBeforeRequestDetails {
   tabId: number;
   url: string;
   type: WebRequestType;
+  requestId: string;
 
   initiator?: string;
   originUrl?: string;
@@ -66,6 +74,111 @@ export function updateResponseHeadersWithCSP(
   return { responseHeaders };
 }
 
+// From https://github.com/kelseasy/web-ext-types/blob/master/global/index.d.ts#L1897
+interface StreamFilter {
+  error: string;
+  status:
+    | 'uninitialized'
+    | 'transferringdata'
+    | 'finishedtransferringdata'
+    | 'suspended'
+    | 'closed'
+    | 'disconnected'
+    | 'failed';
+
+  onstart: (event: any) => void;
+  ondata: (event: { data: ArrayBuffer }) => void;
+  onstop: (event: any) => void;
+  onerror: (event: any) => void;
+
+  close(): void;
+  disconnect(): void;
+  resume(): void;
+  suspend(): void;
+  write(data: Uint8Array | ArrayBuffer): void;
+}
+
+// Detect charset with UTF-8 encoding from HTML
+const CHARSET_TAG_RE = /<meta charset=['"]utf-8/i;
+const CHARSET_HTTP_EQUIV_RE = /<meta http-equiv="content-type" content="text\/html;charset=utf-8/i;
+
+/**
+ * Check if HTML filtering is possible in this browser. Only Firefox is supported.
+ */
+function isHTMLFilteringSupported(): boolean {
+  // @ts-ignore
+  const browser: any = typeof browser !== 'undefined' ? browser : chrome;
+
+  // Apply HTML filtering is any
+  return (
+    typeof TextDecoder !== 'undefined' &&
+    typeof TextEncoder !== 'undefined' &&
+    browser.webRequest !== undefined &&
+    browser.webRequest.filterResponseData !== undefined
+  );
+}
+
+function filterRequestHTML(details: { requestId: string }, rules: HTMLSelector[]): void {
+  // @ts-ignore
+  const browser: any = typeof browser !== 'undefined' ? browser : chrome;
+
+  // Create filter to observe loading of resource
+  const filter: StreamFilter = browser.webRequest.filterResponseData(details.requestId);
+  const decoder = new TextDecoder();
+  const encoder = new TextEncoder();
+  const htmlFilter = new StreamingHtmlFilter(rules);
+  let utf8: undefined | boolean;
+
+  filter.ondata = (event) => {
+    let decoded = '';
+    try {
+      // Attempt decoding chunk (ArrayBuffer) into a string.
+      decoded = decoder.decode(event.data, { stream: true });
+    } catch (ex) {
+      // If we fail to decode chunk, then we need to be extra conservative
+      // and we stop listening to streaming response. This is most likely
+      // because we do not support this encoding.
+      filter.write(event.data);
+      filter.disconnect();
+      return;
+    }
+
+    // Try to guess encoding on first chunk received. The assumption is that
+    // we will find either `charset` or `http-equiv` meta-tag in the header.
+    // If none of this is found, then we fallback to using `isUTF8` which
+    // will make sure `event.data` is valid utf-8 (this check is more
+    // costly).
+    if (utf8 === undefined) {
+      utf8 =
+        CHARSET_TAG_RE.test(decoded) ||
+        CHARSET_HTTP_EQUIV_RE.test(decoded) ||
+        isUTF8(new Uint8Array(event.data));
+    }
+
+    // We only proceed to filter this chunk if we could confirm that encoding
+    // if utf-8. Otherwise we simply proxy the data as-is to not risk
+    // breaking the page.
+    if (utf8 === true) {
+      filter.write(encoder.encode(htmlFilter.write(decoded)));
+    } else {
+      filter.write(event.data);
+    }
+  };
+
+  filter.onstop = async () => {
+    // Make sure we push remaining data if any (needed because both `decoder`
+    // and `htmlFilter` can keep a chunk of data internally which would need
+    // extra input to be processed).
+    const remaining = encoder.encode(htmlFilter.write(decoder.decode())) + htmlFilter.flush();
+
+    if (remaining.length !== 0) {
+      filter.write(encoder.encode(remaining));
+    }
+
+    filter.disconnect();
+  };
+}
+
 /**
  * Wrap `FiltersEngine` into a WebExtension-friendly helper class. It exposes
  * methods to interface with WebExtension APIs needed to block ads.
@@ -112,6 +225,17 @@ export class WebExtensionBlocker extends FiltersEngine {
   ): chrome.webRequest.BlockingResponse => {
     const request = fromWebRequestDetails(details);
     if (request.isMainFrame()) {
+      // Here we optionally perform HTML filtering. This can only be done if:
+      // 1. `enableHtmlFiltering` is set to `true`.
+      // 2. `browser.webRequest.filterResponseData` (Firefox only!).
+      // 3. `TextEncoder` and `TextDecoder` are available.
+      if (this.config.enableHtmlFiltering === true && isHTMLFilteringSupported()) {
+        const htmlFilters = this.getHtmlFilters(request);
+        if (htmlFilters.length !== 0) {
+          filterRequestHTML(details, htmlFilters);
+        }
+      }
+
       return {};
     }
 
@@ -136,7 +260,7 @@ export class WebExtensionBlocker extends FiltersEngine {
       details,
       this.getCSPDirectives(fromWebRequestDetails(details)),
     );
-  }
+  };
 
   private onRuntimeMessage = (
     msg: IBackgroundCallback & { action?: string },
@@ -225,7 +349,7 @@ export class WebExtensionBlocker extends FiltersEngine {
         sendResponse(responseFromBackground);
       }
     }
-  }
+  };
 
   private injectStylesWebExtension(
     styles: string,